added update for 2d objects and loading

+++ Assets/Code/Models/Object2D.cs
@@ -23,7 +23,7 @@ public class Object2D
 
     public float scaleY;
 
-    public Object2D(string environmentId, string prefabId, int positionX, int positionY)
+    public Object2D(string environmentId, string prefabId, float positionX, float positionY)
     {
         this.environmentId = environmentId;
         this.prefabId = prefabId;
